+++ packages/eslint-config-jest/index.js
@@ -1,4 +1,5 @@
 module.exports = {
+  plugins: ['filenames'],
   overrides: [
     {
       files: [
@@ -9,7 +10,11 @@ module.exports = {
         'jest.setup.{ts,js}',
       ],
       excludedFiles: ['cypress/**/*'],
-      extends: ['plugin:jest/recommended', 'plugin:jest/style'],
+      extends: [
+        'plugin:jest/recommended',
+        'plugin:jest/style',
+        'plugin:jest-formatting/recommended',
+      ],
       rules: {
         'jest/prefer-lowercase-title': ['error', { ignore: ['describe'] }],
         'jest/no-disabled-tests': 0,
@@ -20,6 +25,16 @@ module.exports = {
       rules: {
         'jest/require-top-level-describe': 0,
         'import/no-extraneous-dependencies': 0,
+      },
+    },
+    /**
+     * Enforce `.spec` over `.test` file names
+     * {@link https://github.com/reside-eng/guidelines/blob/main/rfcs/ui-patterns/002-unit-test-files.md}
+     */
+    {
+      files: ['**/*.test.*'],
+      rules: {
+        'filenames/match-regex': ['warn', '\\.spec\\.'],
       },
     },
   ],