+++ packages/eslint-config-typescript/index.js
@@ -11,13 +11,13 @@ module.exports = {
     {
       files: ['**/*.ts?(x)'],
       extends: [
-        'plugin:@typescript-eslint/recommended',
-        'plugin:@typescript-eslint/recommended-requiring-type-checking',
+        'plugin:@typescript-eslint/recommended-type-checked',
+        'plugin:@typescript-eslint/stylistic-type-checked',
       ],
       parser: '@typescript-eslint/parser',
       parserOptions: {
         tsconfigRootDir: process.cwd(),
-        project: ['./tsconfig.json'],
+        project: true,
       },
     },
   ],