+++ packages/jest-config-next/.eslintrc.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: ['@side/base', '@side/typescript', 'prettier'],
+  rules: {
+    'no-console': 'off',
+  },
+};