+++ packages/eslint-config-typescript/CHANGELOG.md
@@ -3,6 +3,108 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.2.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.1.0...@side/eslint-config-typescript@1.2.0) (2023-10-24)
+
+### Features
+
+- update typescript config ([#562](https://github.com/reside-eng/lint-config/issues/562)) ([0a814d7](https://github.com/reside-eng/lint-config/commit/0a814d7ac0513db129556aa2cf206f241955e855))
+
+# [1.1.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.19...@side/eslint-config-typescript@1.1.0) (2023-03-29)
+
+### Features
+
+- upgrade lerna ([#492](https://github.com/reside-eng/lint-config/issues/492)) ([54c907b](https://github.com/reside-eng/lint-config/commit/54c907bc7d6dee491fbac3ecb9769284554785cf)), closes [#488](https://github.com/reside-eng/lint-config/issues/488)
+
+# [1.1.0-alpha.0](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.19...@side/eslint-config-typescript@1.1.0-alpha.0) (2023-03-29)
+
+### Features
+
+- upgrade lerna ([#488](https://github.com/reside-eng/lint-config/issues/488)) ([6b10c5e](https://github.com/reside-eng/lint-config/commit/6b10c5ea54a9e68f7e3b04499a48a85b704a93cc))
+
+## [1.0.19](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.18...@side/eslint-config-typescript@1.0.19) (2023-03-13)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#468](https://github.com/reside-eng/lint-config/issues/468)) ([2ab68cd](https://github.com/reside-eng/lint-config/commit/2ab68cd8a2e8cc06899a50beccd99f2b3c9e8d7a))
+
+## [1.0.18](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.17...@side/eslint-config-typescript@1.0.18) (2023-02-27)
+
+### Bug Fixes
+
+- **deps:** update dependency eslint-import-resolver-typescript to v3 ([#384](https://github.com/reside-eng/lint-config/issues/384)) ([5a3d969](https://github.com/reside-eng/lint-config/commit/5a3d969d5e60ca74aaaa2c3fea633e320fbc2946))
+
+## [1.0.17](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.16...@side/eslint-config-typescript@1.0.17) (2022-10-24)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release to v5.40.1 ([#427](https://github.com/reside-eng/lint-config/issues/427)) ([e16e642](https://github.com/reside-eng/lint-config/commit/e16e64297fa7882a035bf25f14781f1e77c93a4e))
+
+## [1.0.16](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.15...@side/eslint-config-typescript@1.0.16) (2022-10-17)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release to v5.40.0 ([#423](https://github.com/reside-eng/lint-config/issues/423)) ([9019371](https://github.com/reside-eng/lint-config/commit/901937106674c12658eeee0c4b0890b9a65a7a7f))
+
+## [1.0.15](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.14...@side/eslint-config-typescript@1.0.15) (2022-10-10)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release to v5.39.0 ([#419](https://github.com/reside-eng/lint-config/issues/419)) ([c63c58c](https://github.com/reside-eng/lint-config/commit/c63c58cc4d30c5bcf6cce7a66161d0b6f8b8e1eb))
+
+## [1.0.14](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.13...@side/eslint-config-typescript@1.0.14) (2022-10-03)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#415](https://github.com/reside-eng/lint-config/issues/415)) ([2e6788f](https://github.com/reside-eng/lint-config/commit/2e6788f1d1a99b018345544bdef690a3faf13170))
+
+## [1.0.13](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.12...@side/eslint-config-typescript@1.0.13) (2022-09-19)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#411](https://github.com/reside-eng/lint-config/issues/411)) ([2b1935a](https://github.com/reside-eng/lint-config/commit/2b1935a8f6a5a0f70df179d30b70d75c3f6fc6d7))
+
+## [1.0.12](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.11...@side/eslint-config-typescript@1.0.12) (2022-08-22)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release to v5.33.1 ([#402](https://github.com/reside-eng/lint-config/issues/402)) ([1de6862](https://github.com/reside-eng/lint-config/commit/1de6862337736a761e6906f03f80e1cb5836215f))
+
+## [1.0.11](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.10...@side/eslint-config-typescript@1.0.11) (2022-08-15)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#400](https://github.com/reside-eng/lint-config/issues/400)) ([1f835aa](https://github.com/reside-eng/lint-config/commit/1f835aaefc0627c70f2258adf5de9612a1186a9e))
+
+## [1.0.10](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.9...@side/eslint-config-typescript@1.0.10) (2022-08-08)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#398](https://github.com/reside-eng/lint-config/issues/398)) ([1a720f1](https://github.com/reside-eng/lint-config/commit/1a720f18e14545d58149ad2c8b63c4e12236e005))
+
+## [1.0.9](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.8...@side/eslint-config-typescript@1.0.9) (2022-08-01)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release to v5.31.0 ([#396](https://github.com/reside-eng/lint-config/issues/396)) ([ebe7f1f](https://github.com/reside-eng/lint-config/commit/ebe7f1f1f84eaf4e88e72f57f26d2edc531d3908))
+
+## [1.0.8](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.7...@side/eslint-config-typescript@1.0.8) (2022-07-25)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#392](https://github.com/reside-eng/lint-config/issues/392)) ([a701d43](https://github.com/reside-eng/lint-config/commit/a701d432f93620b3f3c3e36ba4618f25523f4b02))
+
+## [1.0.7](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.6...@side/eslint-config-typescript@1.0.7) (2022-07-18)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#388](https://github.com/reside-eng/lint-config/issues/388)) ([fe25d35](https://github.com/reside-eng/lint-config/commit/fe25d356faea911c7a59526d503df984c9b92443))
+
+## [1.0.6](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.5...@side/eslint-config-typescript@1.0.6) (2022-07-07)
+
+### Bug Fixes
+
+- **deps:** weekly npm maintenance release ([#381](https://github.com/reside-eng/lint-config/issues/381)) ([5e07237](https://github.com/reside-eng/lint-config/commit/5e07237d9942466e21d189facaf39187025996d9))
+
 ## [1.0.5](https://github.com/reside-eng/lint-config/compare/@side/eslint-config-typescript@1.0.4...@side/eslint-config-typescript@1.0.5) (2022-06-09)
 
 ### Bug Fixes